07/08 Class

+++ Source/Engine/Core/Random.h
@@ -8,7 +8,7 @@ namespace Cpain {
 		inline int getRandomInt(int max) { return rand() % max; }
 		inline int getRandomInt(int min, int max) { return min + getRandomInt(max - min + 1); }
 
-		inline float getRandomFloat() { return rand() / RAND_MAX; }
+		inline float getRandomFloat() { return rand() / static_cast<float>(RAND_MAX); }
 
 	}
